requirements-windows.txt
# Core web framework
flask==2.3.3
werkzeug==2.3.7

# Database
pymongo==4.5.0
dnspython==2.4.2
flask-pymongo==2.3.0

# HTTP and networking
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4

# Text processing and AI/ML
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
beautifulsoup4==4.12.2
trafilatura==1.6.4

# Image processing (Windows-compatible version)
pillow==9.5.0

# Security and authentication
bcrypt==4.0.1
passlib==1.7.4
oauthlib==3.2.2
pyjwt==2.8.0
cryptography==41.0.7

# Email validation
email-validator==2.1.0

# Additional utilities
python-dateutil==2.8.2
six==1.16.0
jsonschema==4.19.0